Refactor list view components
